modularize Engine in order to define other kinds of Engines (#235)
